Added deleteUser, moved getUserChannels from channelController/routes to userController/routes for organization purposes
