+++ routes/channelRoutes.js
@@ -4,6 +4,4 @@ const router = express.Router();
 
 router.route("/").post(channelsController.createChannel);
 
-router.route("/user/:id").get(channelsController.getUserChannels);
-
 module.exports = router;
